Decimal numbers are used now instead of float. Managed to have the result being rounded to two decimal places.

+++ ConversionTable.cs
@@ -24,8 +24,8 @@ namespace CurrencyConverter //Program picks conversions from this class.
 
      
         
-        public static float euroDollar = 1.12f; //currentish exchange rate 
-        public static float dollarEuro = 0.89f; //currentish exchange rate 
+        public static decimal euroDollar = (Decimal)1.12f; //currentish exchange rate 
+        public static decimal dollarEuro = (Decimal)0.89f; //currentish exchange rate 
 
     }
 }
